pyproject.toml
[project]
name = "py-flagsmith-cli"
dynamic = ["version"]
description = "flagsmith-cli Python Implementation."
authors = [{ name = "belingud", email = "im.victor@qq.com" }]
dependencies = ["typer>=0.12.3", "requests>=2.32.3", "python-dotenv>=1.0.1"]
requires-python = "<4.0,>=3.8"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
pysmith = "py_flagsmith_cli.cli:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
addopts = "-rsxX -l --tb=short --strict"
testpaths = ["tests/"]
python_files = ["tests/*/test*.py"]
[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["py_flagsmith_cli"]

[tool.pdm.version]
source = "file"
path = "py_flagsmith_cli/__init__.py"

[tool.pdm.scripts]
format = "black py_flagsmith_cli"
isort = "isort ."
cli = "python -m py_flagsmith_cli"

[tool.pdm.dev-dependencies]
dev = ["tox-pdm>=0.7.2", "pytest>=8.2.1", "isort>=5.13.2", "mock>=5.1.0"]
